+++ package.json
@@ -3,6 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.2.27",
+    "@fortawesome/free-solid-svg-icons": "^5.12.1",
+    "@fortawesome/react-fontawesome": "^0.1.9",
     "@material-ui/core": "^4.9.5",
     "@material-ui/icons": "^4.9.1",
     "bootstrap": "^4.3.1",
@@ -13,7 +16,7 @@
     "react-mic": "^12.4.1",
     "react-particles-js": "^2.7.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.0.1",
+    "react-scripts": "^3.4.0",
     "styled-components": "^4.4.1",
     "universal-cookie": "^4.0.2"
   },
